chore: split context files to fix HMR issues
