+++ shared/src/containers/ProjectTreeTable/context/index.ts
@@ -2,10 +2,18 @@ export * from './clipboard'
 
 export * from './CellEditingContext'
 export * from './ClipboardContext'
+// column settings context
+export * from './ColumnSettingsProvider'
 export * from './ColumnSettingsContext'
+// project table context
 export * from './ProjectTableContext'
+export * from './ProjectTableProvider'
 export * from './ProjectTableQueriesContext'
+// selected rows context
 export * from './SelectedRowsContext'
+export * from './SelectedRowsProvider'
+// selection cells context
 export * from './SelectionCellsContext'
-export * from './ProjectTableSelectionContext'
+export * from './SelectionCellsProvider'
+// full table selection context
 export * from './ProjectDataContext'
