+++ src/pages/ProjectPage/ProjectPage.tsx
@@ -18,7 +18,7 @@ import { useGetProjectAddonsQuery, useListAddonsQuery } from '@shared/api'
 import { TabPanel, TabView } from 'primereact/tabview'
 import AppNavLinks from '@containers/header/AppNavLinks'
 import { SlicerProvider } from '@context/SlicerContext'
-import { EntityListsProvider } from '@pages/ProjectListsPage/context/EntityListsContext'
+import { EntityListsProvider } from '@pages/ProjectListsPage/context'
 import useLoadRemoteProjectPages from '../../remote/useLoadRemotePages'
 import { Navigate } from 'react-router-dom'
 import ProjectPubSub from './ProjectPubSub'
